requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
apscheduler
sqlalchemy
psycopg2-binary==2.9.10